+++ src/cli.rs
@@ -69,7 +69,7 @@ pub fn main() -> Result<()> {
                 .global(true)
                 .num_args(1)
                 .action(ArgAction::Append)
-                .help("Suppresses specific warnings.")
+                .help("Suppresses specific warnings. Use `all` to suppress all warnings.")
                 .value_parser(value_parser!(String)),
         )
         .subcommand(cmd::update::new())
@@ -110,6 +110,12 @@ pub fn main() -> Result<()> {
         .map(|s| s.to_owned())
         .collect();
 
+    let suppressed_warnings: IndexSet<String> = if suppressed_warnings.contains("all") {
+        (1..18).map(|i| format!("W{:02}", i)).collect()
+    } else {
+        suppressed_warnings
+    };
+
     // Enable debug outputs if needed.
     if matches.contains_id("debug") && matches.get_flag("debug") {
         ENABLE_DEBUG.store(true, std::sync::atomic::Ordering::Relaxed);
@@ -143,13 +149,14 @@ pub fn main() -> Result<()> {
 
     // Parse the manifest file of the package.
     let manifest_path = root_dir.join("Bender.yml");
-    let manifest = read_manifest(&manifest_path)?;
+    let manifest = read_manifest(&manifest_path, &suppressed_warnings)?;
     debugln!("main: {:#?}", manifest);
 
     // Gather and parse the tool configuration.
     let config = load_config(
         &root_dir,
         matches!(matches.subcommand(), Some(("update", _))) && !suppressed_warnings.contains("W02"),
+        &suppressed_warnings,
     )?;
     debugln!("main: {:#?}", config);
 
@@ -376,7 +383,7 @@ fn find_package_root(from: &Path) -> Result<PathBuf> {
 }
 
 /// Read a package manifest from a file.
-pub fn read_manifest(path: &Path) -> Result<Manifest> {
+pub fn read_manifest(path: &Path, suppress_warnings: &IndexSet<String>) -> Result<Manifest> {
     use crate::config::PartialManifest;
     use std::fs::File;
     debugln!("read_manifest: {:?}", path);
@@ -387,12 +394,16 @@ pub fn read_manifest(path: &Path) -> Result<Manifest> {
     partial
         .prefix_paths(path.parent().unwrap())
         .map_err(|cause| Error::chain(format!("Error in manifest prefixing {:?}.", path), cause))?
-        .validate("", false)
+        .validate("", false, suppress_warnings)
         .map_err(|cause| Error::chain(format!("Error in manifest {:?}.", path), cause))
 }
 
 /// Load a configuration by traversing a directory hierarchy upwards.
-fn load_config(from: &Path, warn_config_loaded: bool) -> Result<Config> {
+fn load_config(
+    from: &Path,
+    warn_config_loaded: bool,
+    suppress_warnings: &IndexSet<String>,
+) -> Result<Config> {
     #[cfg(unix)]
     use std::os::unix::fs::MetadataExt;
 
@@ -465,7 +476,7 @@ fn load_config(from: &Path, warn_config_loaded: bool) -> Result<Config> {
 
     // Validate the configuration.
     let mut out = out
-        .validate("", false)
+        .validate("", false, suppress_warnings)
         .map_err(|cause| Error::chain("Invalid configuration:", cause))?;
 
     out.overrides = out