+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add check to ensure files referenced in all manifests exist.
 - Add warnings for unknown fields in manifest.
 - Add support to indicate and read in external file lists in manifest.
+- Enumerate warnings and add `--suppress` flag to hide warnings.
 
 ### Changed
 - `update`: Clean up alignment of manual resolution output.